requirements.txt
langchain-google-genai
google-generativeai==0.8.3
langchain-openai
langgraph
fastapi
uvicorn
python-dotenv
pydantic
